# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API Barantin",
        "description": "Data Request",
        "contact": {
            "email": "nandaputraservice@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Api Admin Barantin"
        },
        {
            "url": "http://localhost:8000/api/v2",
            "description": "Api User Barantin Test"
        }
    ],
    "paths": {
...
        "/user/mitra": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Ambil Semua User Mitra di masing masing perngguna jasa",
                "operationId": "c84f56f1eedec5e96f85188a1f2ad418",
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "nama_mitra": {
                                                        "type": "string",
                                                        "example": "Nama Mitra"
                                                    },
                                                    "jenis_identitas_mitra": {
                                                        "type": "string",
                                                        "example": "KTP"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/user/mitra/{mitra_id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Ambil data mitra pengguna jasa berdasarkan id",
                "operationId": "aa63ef6d1aeff856fa4b5bfe0d43af92",
                "parameters": [
                    {
                        "in": "path",
                        "name": "mitra_id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "nama_mitra": {
                                                    "type": "string",
                                                    "example": "Nama Mitra"
                                                },
                                                "jenis_identitas_mitra": {
                                                    "type": "string",
                                                    "example": "KTP"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/user/ppjk": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get PPJK",
                "description": "Get PPJK Pengguna Jasa",
                "operationId": "getPpjk",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Data Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/user/ppjk/{ppjk_id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get Detail PPJK Pengguna Jasa",
                "description": "Get Detail PPJK",
                "operationId": "getDetailPpjk",
                "parameters": [
                    {
                        "name": "ppjk_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Data Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/user/profile": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get User Profile",
                "description": "Get User Profile",
                "operationId": "bfb90b10e3a49ea12c6614eb9db0ef95",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Data Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Tambahkan Kata \"Bearer\" Sebelum token contoh: (Bearer xx|497zSWISvECINqxxxxxxxxxxxxxxx)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}